refactor map with util function
